+++ lib/puffer_pages/liquid/tags/helper.rb
@@ -1,0 +1,15 @@
+module PufferPages
+  module Liquid
+    module Tags
+
+      class Helper < ::Liquid::Tag
+        def render(context)
+          context.registers[:tracker].register("<%= #{@tag_name} %>")
+        end
+      end
+
+    end
+  end
+end
+
+Liquid::Template.register_tag('csrf_meta_tags', PufferPages::Liquid::Tags::Helper)