+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### New features
 
+*   `{% csrf_meta_tags %}` helper tag.
+
 *   The new `puffer_pages` method for controller
 
     ```
@@ -21,7 +23,7 @@
 
 *   The only way to declare puffer_page rendering
     `render :puffer_page => Page.first` or
-    `render :puffer_page => 'some/path'`
+    `render :puffer_page => 'some/path'`.
 
 *   Current controller now available from context.registers[:controller]
     It is useful for using url helpers in drops.